+++ base/CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 - **BREAKING**: Replace `error-chain` with `failure`.
 - Standardise API error response.
+- Standardise logging across core and agents.
 
 ## [0.2.0] - 2019-02-20
 ### Added