+++ base/CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.3.0] - 2019-03-29
 ### Changed
 - **BREAKING**: Replace `error-chain` with `failure`.
 - Standardise API error response.
@@ -31,5 +33,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - OpenTracing integration
 
 
-[Unreleased]: https://github.com/replicante-io/agents/compare/v0.2.0...HEAD
+[Unreleased]: https://github.com/replicante-io/agents/compare/v0.3.0...HEAD
+[0.3.0]: https://github.com/replicante-io/agents/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/replicante-io/agents/compare/v0.1.0...v0.2.0